Keep matlab functions together
