+++ @amimodel/compileC.m
@@ -420,14 +420,7 @@ function compileC(this)
             fullfile(this.wrap_path,'src','amici.cpp') ' ' ...
             includesstr]);
     end
-    
-    if(this.nxtrue == this.nx)
-        eval(['mex ' DEBUG ' ' COPT ...
-            ' -c -outdir ' fullfile(this.wrap_path,'src') ' ' ...
-            fullfile(this.wrap_path,'src','amiwrap.cpp') ' ' ...
-            includesstr]);
-    end
-    
+        
     if(this.nxtrue == this.nx)
         eval(['mex ' DEBUG ' ' COPT ...
             ' -c -outdir ' fullfile(this.wrap_path,'src') ' ' ...
@@ -437,7 +430,6 @@ function compileC(this)
 
     eval(['mex ' DEBUG ' ' COPT ' ' CLIBS ...
         ' -output ' fullfile(this.wrap_path,'models',this.modelname,['ami_' this.modelname]) ...
-        ' "' fullfile(this.wrap_path,'src',['amiwrap' o_suffix]) '"' ...
         ' "' fullfile(this.wrap_path,'src',['amici' o_suffix]) '"' ...
         ' "' fullfile(this.wrap_path,'src',['amici_interface_matlab' o_suffix]) '"' ...
         ' "' fullfile(this.wrap_path,'models',this.modelname,['wrapfunctions' o_suffix]) '"' ...
