+++ src/amici_interface_matlab.cpp
@@ -1,5 +1,11 @@
-#include "include/amici_interface_matlab.h"
+/**
+ * @file   amiwrap.cpp
+ * @brief  core routines for mex interface
+ *
+ * This file defines the fuction mexFunction which is executed upon calling the mex file from matlab
+ */
 
+#include "include/amici_interface_matlab.h"
 #include "wrapfunctions.h" /* user functions */
 
 #include <cstring>
@@ -7,6 +13,12 @@
 #include <blas.h>
 #include <include/edata_accessors.h>
 #include <include/tdata_accessors.h>
+
+#define _USE_MATH_DEFINES /* MS definition of PI and other constants */
+#include <cmath>
+#ifndef M_PI /* define PI if we still have no definition */
+#define M_PI 3.14159265358979323846
+#endif
 
 /**
  * @ brief extract information from a property of a matlab class (scalar)
@@ -36,6 +48,35 @@
         warnMsgIdAndTxt("AMICI:mex:OPTION","Provided options do not have field " #OPTION "!"); \
         return(udata); \
     }
+
+
+
+/*!
+ * mexFunction is the main function of the mex simulation file this function carries out all numerical integration and writes results into the sol struct.
+ *
+ * @param[in] nlhs number of output arguments of the matlab call @type int
+ * @param[out] plhs pointer to the array of output arguments @type mxArray
+ * @param[in] nrhs number of input arguments of the matlab call @type int
+ * @param[in] prhs pointer to the array of input arguments @type mxArray
+ * @return void
+ */
+void mexFunction(int nlhs, mxArray *plhs[], int nrhs, const mxArray *prhs[]) {
+    /* return status flag */
+    int status = 0;
+
+    UserData udata = userDataFromMatlabCall(prhs, &status);
+    ReturnDataMatlab rdata(&udata);
+
+    plhs[0] = rdata.mxsol;
+
+    if(status == 0 && udata.nx > 0) {
+        ExpData edata = expDataFromMatlabCall(prhs, &udata, &status);
+        if (status == 0)
+            runAmiciSimulation(&udata, &edata, &rdata, &status);
+    }
+
+    *rdata.status = (double) status;
+}
 
 
 UserData userDataFromMatlabCall(const mxArray *prhs[], int *status) {
